neptune.ai/sandbox/neptune-iris/requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
bravado==11.0.2
bravado-core==5.17.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.13
idna==2.10
importlib-metadata==3.7.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonpointer==2.0
jsonref==0.2
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
MarkupSafe==1.1.1
mistune==0.8.4
monotonic==1.5
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
neptune-client==0.4.132
neptune-notebooks==0.0.19
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.20.1
oauthlib==3.1.0
packaging==20.9
pandas==1.2.2
pandocfilters==1.4.3
parso==0.8.1
path==15.1.2
path.py==12.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.0
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rfc3987==1.3.8
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
smmap==3.0.5
strict-rfc3339==0.7
swagger-spec-validator==2.7.3
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
websocket-client==0.57.0
zipp==3.4.0
